+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Yii Validator Change Log
 
-## 1.2.0 under development
+## 2.0.0 under development
 
 - New #597, #608: Add debug collector for `yiisoft/yii-debug` (@xepozz, @vjik)
 - New #610: Add `$escape` parameter to methods `Result::getAttributeErrorMessagesIndexedByPath()` and
@@ -8,6 +8,9 @@
   elements (@vjik)
 - Bug #612: Disable escaping of asterisk char in value path returned by `Error::getValuePath(true)` (@vjik)
 - New #617: Add `OneOf` rule (@arogachev)
+- Chg #623: List translated attributes in error message for `OneOf` and `AtLeast` rules (@arogachev)
+- Chg #624: Fix meaning of error message in `OneOf` rule (@arogachev)
+- Chg #625: Improve meaning and use pluralization in error message for `OneOf` and `AtLeast` rules (@arogachev)
 
 ## 1.1.0 April 06, 2023